+++ fmt.go
@@ -5,7 +5,7 @@ import (
 )
 
 func (sq *Sq) Scan(state fmt.ScanState, _ rune) error {
-	var file, rank byte
+	var file, rank int16
 	r, _, err := state.ReadRune()
 	if err != nil {
 		return err
@@ -15,7 +15,7 @@ func (sq *Sq) Scan(state fmt.ScanState, _ rune) error {
 	} else if 'a' > r || r > 'h' {
 		return fmt.Errorf("invalid character for square's file: %c", r)
 	}
-	file = uint8(r - 'a')
+	file = int16(r - 'a')
 
 	r, _, err = state.ReadRune()
 	if err != nil {
@@ -23,7 +23,7 @@ func (sq *Sq) Scan(state fmt.ScanState, _ rune) error {
 	} else if '1' > r || r > '8' {
 		return fmt.Errorf("invalid character for square's rank: %c", r)
 	}
-	rank = uint8(r-'1') << 4
+	rank = int16(r-'1') << 4
 
 	*sq = Sq(rank | file)
 	return nil
